pyproject.toml
[tool.poetry]
name = "docusketchtest-fastapi"
version = "0.1.0"
description = "Fast Api test task"
authors = ["Your Name <you@example.com>"]
license = "BSD"
readme = "README.md"
packages = [{include = "docusketchtest_fastapi"}]

[tool.poetry.dependencies]
python = "3.11.0"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-decouple = "^3.8"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
psycopg2 = "^2.9.7"
pandas = "^2.1.0"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
